extract line editor
